requirements.txt
discord.py
mecha